implement JSONserialization category
