--- Artist-objc/Artist-objcTests/Artist_objcTests.m
@@ -1,37 +1,0 @@
-//
-//  Artist_objcTests.m
-//  Artist-objcTests
-//
-//  Created by Bradley Yin on 10/11/19.
-//  Copyright © 2019 bradleyyin. All rights reserved.
-//
-
-#import <XCTest/XCTest.h>
-
-@interface Artist_objcTests : XCTestCase
-
-@end
-
-@implementation Artist_objcTests
-
-- (void)setUp {
-    // Put setup code here. This method is called before the invocation of each test method in the class.
-}
-
-- (void)tearDown {
-    // Put teardown code here. This method is called after the invocation of each test method in the class.
-}
-
-- (void)testExample {
-    // This is an example of a functional test case.
-    // Use XCTAssert and related functions to verify your tests produce the correct results.
-}
-
-- (void)testPerformanceExample {
-    // This is an example of a performance test case.
-    [self measureBlock:^{
-        // Put the code you want to measure the time of here.
-    }];
-}
-
-@end
