+++ Artist-objc/Artist-objcTests/BYArtistTest.m
@@ -1,0 +1,20 @@
+//
+//  BYArtistTest.m
+//  Artist-objcTests
+//
+//  Created by Bradley Yin on 10/11/19.
+//  Copyright © 2019 bradleyyin. All rights reserved.
+//
+
+#import <XCTest/XCTest.h>
+#import "BYFileHelper.h"
+#import "BYArtist.h"
+#import "BYArtist+BYNSJSONSerialization"
+
+@interface BYArtistTest : XCTestCase
+
+@end
+
+@implementation BYArtistTest
+
+@end
